Remove 'gpu' label from node_totals metric
We can actually replicate this functionality by just using node_type
regex queries in Prometheus, which is more TheWay(tm) to do this.

Also moves the gauge declaration out; the .labels() function creates a
new context that the value will be set on. With the gauge being
redefined in each loop, all the Help text and stuff gets re-added to the
output each time.